requirements.txt
Flask
requests
gTTS
SpeechRecognition
pydub
pdfplumber
pytesseract
Pillow
gunicorn
python-dotenv
